+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Global binaries can now be executed such as `nve 8 npm test`. Keep in mind
   that the binary must be compatible with the chosen Node.js version. For
   example `npm` is only compatible with Node `>=6`.
+- Local binaries can be executed as well
 - If the script spawn child processes, those will now re-use the same Node.js
   version
 - Add